Build 50 (Pre-Xcode 10 Beta)
Glow v1.0, Build 50
The development for the Glow project will be moving forward in Xcode 10, Swift 4.2 beta territory to prepare for an expected post-Mojave launch. I want to ensure that Glow will be compatible with the latest platforms. Furthermore, with macOS 10.14, Glow has the ability to access many more features that were not possible before.